[feat]: app.js파일 분할
