+++ src/App.test.js
@@ -1,6 +1,5 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from './App1';
 
 test('renders learn react link', () => {
   render(<App />);
